Several important bug fixes. See description.
Made the Snowflake Shuriken fire a projectile, changed the entity
registry methods to use the FML mod entity system, allowing DivineRPG
to be used with other mods that add entities (it was previously filling
up all the available slots). Changed the entity names to not contain
spaces, as these were preventing the entities from being spawned with
the /summon command. Also, updated the armor bonus field mappings to
1.7.10 and fixed an angelic armor bug that prevented flying without the
armor in creative.