Fractite cannon shooting, projectile, and renderer, a few more fixes.
Serenade striker now only works with 100 blocks of the target (as in
the previous version). As I could not find a texture for the fracture
cannon projectile, I made one. Also, made the Frossivence function and
fixed the healing amount of the Massivence (should be two health, not
one). More work on fixing #16 and registering items in preinit,
manually, like mods should do.