Update zh_cn.lang for 1.12.2 and a small request
This is an update for zh_cn.lang (1.12.2).  And I'm wondering if I can repost DivineRPG to a Chinese forum called MCBBS, you know, because of GFW , it's hard for Chinese players to know many other mods. And I will not use is for profit and I will link to curseforge pages.I will keep everything you have asked.